+++ src/CACHE_VERSION/index.d.ts
@@ -1,0 +1,10 @@
+import type { be_atom_triple_T, WritableAtom_ } from '@ctx-core/nanostores'
+export declare const [
+	CACHE_VERSION$_,
+	CACHE_VERSION_,
+	CACHE_VERSION__set,
+]:be_atom_triple_T<WritableAtom_<false|string>>
+export {
+	CACHE_VERSION$_ as CACHE_VERSION__,
+	CACHE_VERSION$_ as b__CACHE_VERSION,
+}