+++ src/CACHE_VERSION__/index.d.ts
@@ -1,7 +1,0 @@
-import type { WritableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-export declare const CACHE_VERSION__:Be<WritableAtom_<false|string>>
-export {
-	CACHE_VERSION__ as CACHE_VERSION$_,
-	CACHE_VERSION__ as b__CACHE_VERSION,
-}