+++ src/NODE_ENV/index.d.ts
@@ -1,0 +1,10 @@
+import type { be_atom_triple_T, WritableAtom_ } from '@ctx-core/nanostores'
+export declare const [
+	NODE_ENV$_,
+	NODE_ENV_,
+	NODE_ENV__set,
+]:be_atom_triple_T<WritableAtom_<string>>
+export {
+	NODE_ENV$_ as NODE_ENV__,
+	NODE_ENV$_ as b__NODE_ENV,
+}