+++ src/NODE_ENV__/index.d.ts
@@ -1,7 +1,0 @@
-import type { WritableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-export declare const NODE_ENV__:Be<WritableAtom_<string>>
-export {
-	NODE_ENV__ as NODE_ENV$_,
-	NODE_ENV__ as b__NODE_ENV,
-}