+++ src/VERSION/index.d.ts
@@ -1,0 +1,7 @@
+import type { be_atom_triple_T, WritableAtom_ } from '@ctx-core/nanostores'
+export declare const [
+	VERSION$_,
+	VERSION_,
+	VERSION__set,
+]:be_atom_triple_T<WritableAtom_<string>>
+export { VERSION$_ as VERSION__ }