+++ src/VERSION__/index.d.ts
@@ -1,4 +1,0 @@
-import type { WritableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-export declare const VERSION__:Be<WritableAtom_<string>>
-export { VERSION__ as VERSION$_ }