+++ src/is_development/index.d.ts
@@ -1,0 +1,6 @@
+import type { be_computed_pair_T, ReadableAtom_ } from '@ctx-core/nanostores'
+export declare const [
+	is_development$_,
+	is_development_,
+]:be_computed_pair_T<ReadableAtom_<boolean>>
+export { is_development$_ as is_development__ }