+++ src/is_development__/index.d.ts
@@ -1,4 +1,0 @@
-import type { ReadableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-export declare const is_development__:Be<ReadableAtom_<boolean>>
-export { is_development__ as is_development$_ }