+++ src/missing_env__throw/index.js
@@ -3,7 +3,7 @@ import { error_o__throw } from '@ctx-core/error'
  * Throws an error for a missing env variable
  */
 /** @type {import('missing_env__throw').missing_env__throw} */
-export const missing_env__throw = (env_name)=>{
+export const missing_env__throw = env_name=>{
 	const error_message = `
 ${env_name} environment variable not set.
 dev: make sure ${env_name} is set in your .env file
@@ -16,5 +16,5 @@ heroku: make sure ${env_name} is set using \`heroku config:set\`
 }
 export {
 	missing_env__throw as throw_missing_env,
-	missing_env__throw as throw__missing__env, 
+	missing_env__throw as throw__missing__env,
 }